Move schema to own folder
